simple page with header only
This is imperative code style. We provide steps for how the UI should be
updated. For this simple header we had multiple locs to write.
This is not scalable for big apps.

React is a popular declarative library that solves the above challenge.
Declarative means we say what the UI should look like and then React
will figure out how to update the DOM.