state
use hooks to add additional logic like state. state is information in
the UI that changes over time, usually triggered by user interaction eg
number of times the like button has been clicked.

`useState` returns an array with first item `value` and a function to
update the value. We use array destructuring to get these.

Unlike props which are passed to components as the first function
parameter, state is initiated and stored within a component. You may
pass the state information to children components as props, but the
logic for updating the state should be kept within the component where
state was initially created.

props is read-only information that is passed to components whereas
state is information that can change over time, usually triggered by
user interaction.